# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PyExec Portal API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Department {
  id: number;
  name: string;
  description?: string | null;
  createdAt: string;
}

export interface CreateDepartmentBody {
  name: string;
  description?: string | null;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  admin: "admin",
  user: "user",
} as const;

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  firstName?: string | null;
  lastName?: string | null;
  role: UserProfileRole;
  departmentId?: number | null;
  departmentName?: string | null;
  createdAt: string;
}

export interface UpdateProfileBody {
  firstName?: string;
  lastName?: string;
}

export interface AssignDepartmentBody {
  departmentId: number | null;
}

export type AssignRoleBodyRole =
  (typeof AssignRoleBodyRole)[keyof typeof AssignRoleBodyRole];

export const AssignRoleBodyRole = {
  admin: "admin",
  user: "user",
} as const;

export interface AssignRoleBody {
  role: AssignRoleBodyRole;
}

export interface Script {
  id: number;
  name: string;
  description?: string | null;
  filename: string;
  code: string;
  departmentId?: number | null;
  departmentName?: string | null;
  uploadedBy: string;
  uploadedByName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface UploadScriptBody {
  name: string;
  description?: string | null;
  filename: string;
  code: string;
  departmentId?: number | null;
}

export interface ExecuteScriptBody {
  args?: string[];
  stdin?: string | null;
}

export interface ExecutionResult {
  success: boolean;
  stdout: string;
  stderr: string;
  exitCode: number;
  executionTimeMs: number;
  executionId?: number;
}

export type AuditLogDetails = { [key: string]: unknown } | null;

export interface AuditLog {
  id: number;
  userId?: string | null;
  userEmail?: string | null;
  action: string;
  resourceType?: string | null;
  resourceId?: string | null;
  details?: AuditLogDetails;
  ipAddress?: string | null;
  createdAt: string;
}

export interface AuditLogPage {
  logs: AuditLog[];
  total: number;
  page: number;
  limit: number;
}

export type DashboardStatsScriptsByDepartmentItem = {
  departmentName: string;
  count: number;
};

export interface DashboardStats {
  totalScripts: number;
  totalUsers: number;
  totalDepartments: number;
  totalExecutions: number;
  recentExecutions: number;
  scriptsByDepartment: DashboardStatsScriptsByDepartmentItem[];
  recentActivity: AuditLog[];
}

export type ListAuditLogsParams = {
  page?: number;
  limit?: number;
  action?: string;
  userId?: string;
};