# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PyExec Portal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a new department (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "403":
          description: Forbidden
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get a department by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "404":
          description: Not found
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
...
  /scripts:
    get:
      operationId: listScripts
      tags: [scripts]
      summary: List scripts accessible to current user (filtered by department)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of scripts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Script"
    post:
      operationId: uploadScript
      tags: [scripts]
      summary: Upload a new Python script
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadScriptBody"
      responses:
        "201":
          description: Script uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
  /scripts/{id}:
    get:
      operationId: getScript
      tags: [scripts]
      summary: Get a script by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Script details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
        "403":
          description: Access denied
        "404":
          description: Not found
    delete:
      operationId: deleteScript
      tags: [scripts]
      summary: Delete a script (admin or uploader)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
  /scripts/{id}/execute:
    post:
      operationId: executeScript
      tags: [execution]
      summary: Execute a Python script
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteScriptBody"
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResult"
        "403":
          description: Access denied
  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [audit]
      summary: Get audit logs (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: action
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Audit logs with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogPage"
        "403":
          description: Forbidden
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [audit]
      summary: Get dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
    CreateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
      required:
        - name
...
    Script:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        filename:
          type: string
        code:
          type: string
        departmentId:
          type: integer
          nullable: true
        departmentName:
          type: string
          nullable: true
        uploadedBy:
          type: string
        uploadedByName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - filename
        - code
        - uploadedBy
        - createdAt
        - updatedAt
    UploadScriptBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        filename:
          type: string
        code:
          type: string
        departmentId:
          type: integer
          nullable: true
      required:
        - name
        - filename
        - code
    ExecuteScriptBody:
      type: object
      properties:
        args:
          type: array
          items:
            type: string
        stdin:
          type: string
          nullable: true
    ExecutionResult:
      type: object
      properties:
        success:
          type: boolean
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        executionTimeMs:
          type: integer
        executionId:
          type: integer
      required:
        - success
        - stdout
        - stderr
        - exitCode
        - executionTimeMs
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
          nullable: true
        userEmail:
          type: string
          nullable: true
        action:
          type: string
        resourceType:
          type: string
          nullable: true
        resourceId:
          type: string
          nullable: true
        details:
          type: object
          nullable: true
        ipAddress:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - action
        - createdAt
    AuditLogPage:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - logs
        - total
        - page
        - limit
    DashboardStats:
      type: object
      properties:
        totalScripts:
          type: integer
        totalUsers:
          type: integer
        totalDepartments:
          type: integer
        totalExecutions:
          type: integer
        recentExecutions:
          type: integer
        scriptsByDepartment:
          type: array
          items:
            type: object
            properties:
              departmentName:
                type: string
              count:
                type: integer
            required:
              - departmentName
              - count
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
      required:
        - totalScripts
        - totalUsers
        - totalDepartments
        - totalExecutions
        - recentExecutions
        - scriptsByDepartment
        - recentActivity