# lib/api-zod/src/generated/api/departments/departments.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PyExec Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary List all departments
 */
export const ListDepartmentsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListDepartmentsResponse = zod.array(ListDepartmentsResponseItem);

/**
 * @summary Create a new department (admin only)
 */
export const CreateDepartmentBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
});

/**
 * @summary Get a department by ID
 */
export const GetDepartmentParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDepartmentResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a department (admin only)
 */
export const DeleteDepartmentParams = zod.object({
  id: zod.coerce.number(),
});