# lib/api-zod/src/generated/api/execution/execution.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PyExec Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Execute a Python script
 */
export const ExecuteScriptParams = zod.object({
  id: zod.coerce.number(),
});

export const ExecuteScriptBody = zod.object({
  args: zod.array(zod.string()).optional(),
  stdin: zod.string().nullish(),
});

export const ExecuteScriptResponse = zod.object({
  success: zod.boolean(),
  stdout: zod.string(),
  stderr: zod.string(),
  exitCode: zod.number(),
  executionTimeMs: zod.number(),
  executionId: zod.number().optional(),
});