# lib/api-zod/src/generated/api/scripts/scripts.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PyExec Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary List scripts accessible to current user (filtered by department)
 */
export const ListScriptsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  filename: zod.string(),
  code: zod.string(),
  departmentId: zod.number().nullish(),
  departmentName: zod.string().nullish(),
  uploadedBy: zod.string(),
  uploadedByName: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListScriptsResponse = zod.array(ListScriptsResponseItem);

/**
 * @summary Upload a new Python script
 */
export const UploadScriptBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
  filename: zod.string(),
  code: zod.string(),
  departmentId: zod.number().nullish(),
});

/**
 * @summary Get a script by ID
 */
export const GetScriptParams = zod.object({
  id: zod.coerce.number(),
});

export const GetScriptResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  filename: zod.string(),
  code: zod.string(),
  departmentId: zod.number().nullish(),
  departmentName: zod.string().nullish(),
  uploadedBy: zod.string(),
  uploadedByName: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a script (admin or uploader)
 */
export const DeleteScriptParams = zod.object({
  id: zod.coerce.number(),
});